* Spamd returns marks as double, so parse them as doubles * Fix memory leak (symbols may not be freed) * Change logging style similar to libclamd * Make implementation when system has no sendfile

+++ libspamd.h
@@ -12,6 +12,6 @@
 
 struct config_file;
 
-int spamdscan(const char *file, struct config_file *cfg, int spam_mark[2]);
+int spamdscan(const char *file, struct config_file *cfg, double spam_mark[2]);
 
 #endif
